# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DoD AI Tool Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /library/stats:
    get:
      tags: [Library]
      operationId: getLibraryStats
      summary: Stats for the user's RAG library
      responses:
        '200':
          description: Stats
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LibraryStats' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /library/documents:
    get:
      tags: [Library]
      operationId: listDocuments
      summary: List the user's documents
      parameters:
        - name: presetId
          in: query
          required: false
          description: >
            When set, only return documents currently linked to that mission preset.
          schema: { type: string }
      responses:
        '200':
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DocumentSummary' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    post:
      tags: [Library]
      operationId: uploadTextDocument
      summary: Upload a document by sending its raw text (for plain-text uploads)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TextDocumentUploadRequest' }
      responses:
        '200':
          description: Created document
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentSummary' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /library/auto-ingest:
    post:
      tags: [Library]
      operationId: triggerAutoIngest
      summary: Trigger auto-ingest of a curated MOS or Unit doctrine package
      description: >
        Looks up the curated doctrine package for the given source
        ("mos:<branch>:<code>" or "unit:<branch>:<id>"), downloads each
        public-doctrine URL server-side, extracts and chunks it, and
        inserts the new docs into the user's library. Existing docs with
        the same source URL are skipped, so re-running this is safe and
        idempotent.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AutoIngestRequest' }
      responses:
        '200':
          description: Ingest finished synchronously
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AutoIngestSummary' }
        '400':
          description: Bad request (no curated package for source)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /library/auto-ingest/status:
    get:
      tags: [Library]
      operationId: getAutoIngestStatus
      summary: Poll the most recent auto-ingest job for a source
      parameters:
        - name: source
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Status (or null when no job exists)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AutoIngestStatusResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /library/test-query:
    post:
      tags: [Library]
      operationId: testLibraryQuery
      summary: Run a sample RAG query against the user's own library (dry-run)
      description: >
        Lets the signed-in user verify their library is returning useful
        snippets BEFORE trusting it with a tool launch. Identical retrieval
        path to the launch-token-scoped /tools/library-query endpoint, but
        scoped to the current authenticated user instead of a session token.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LibraryTestQueryRequest' }
      responses:
        '200':
          description: Snippets
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LibraryTestQueryResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /library/documents/{id}:
    get:
      tags: [Library]
      operationId: getDocument
      summary: Get document metadata + first chunks
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Document detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentDetail' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    delete:
      tags: [Library]
      operationId: deleteDocument
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SimpleSuccess' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /library/documents/{id}/presets:
    put:
      tags: [Library]
      operationId: setDocumentPresetTags
      summary: Replace the set of presets a document belongs to
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DocumentPresetTagsUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentSummary' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }

  # ---------- Launch ----------
...
components:
...
  schemas:
...
    SimpleSuccess:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    LibraryStats:
      type: object
      required: [documentCount, chunkCount, totalChars]
      properties:
        documentCount: { type: integer }
        chunkCount: { type: integer }
        totalChars: { type: integer }
    DocumentSummary:
      type: object
      required:
        [id, title, sourceFilename, mimeType, sizeBytes, charCount,
         chunkCount, status, autoSource, sourceUrl, errorMessage,
         uploadedAt, processedAt, presetIds]
      properties:
        id: { type: string }
        title: { type: string }
        sourceFilename: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: integer }
        charCount: { type: integer }
        chunkCount: { type: integer }
        status: { type: string }
        autoSource:
          type: [string, 'null']
          description: >
            Stable identifier of the auto-ingest source that produced this
            doc, e.g. "mos:army:11B" or "unit:marines:MALS-12". Null for
            documents the user uploaded by hand.
        sourceUrl:
          type: [string, 'null']
          description: Original public URL the doc was downloaded from (auto-ingest only).
        errorMessage: { type: [string, 'null'] }
        uploadedAt: { type: string, format: date-time }
        processedAt: { type: [string, 'null'], format: date-time }
        presetIds:
          type: array
          description: IDs of every mission preset this document is currently linked to.
          items: { type: string }
    DocumentDetail:
      allOf:
        - $ref: '#/components/schemas/DocumentSummary'
        - type: object
          required: [chunks]
          properties:
            chunks:
              type: array
              items: { $ref: '#/components/schemas/ChunkPreview' }
    ChunkPreview:
      type: object
      required: [id, chunkIndex, content, charCount]
      properties:
        id: { type: string }
        chunkIndex: { type: integer }
        content: { type: string }
        charCount: { type: integer }
    TextDocumentUploadRequest:
      type: object
      required: [title, sourceFilename]
      properties:
        title: { type: string, minLength: 1 }
        sourceFilename: { type: string, minLength: 1 }
        mimeType: { type: string }
        content:
          type: string
          description: >
            UTF-8 plain-text body. Set this for paste-text uploads. Either
            content or contentBase64 must be provided.
        contentBase64:
          type: string
          description: >
            Base64-encoded binary body for PDF/DOCX/MD/TXT uploads. The server
            picks the right extractor based on mimeType. Either content or
            contentBase64 must be provided.
    LibraryTestQueryRequest:
      type: object
      required: [query]
      properties:
        query: { type: string, minLength: 1 }
        limit: { type: integer }
    LibraryTestQueryResponse:
      type: object
      required: [query, snippets]
      properties:
        query: { type: string }
        snippets:
          type: array
          items: { $ref: '#/components/schemas/RagSnippet' }
...
    RagSnippet:
      type: object
      required:
        [chunkId, documentId, documentTitle, chunkIndex, content, score]
      properties:
        chunkId: { type: string }
        documentId: { type: string }
        documentTitle: { type: string }
        chunkIndex: { type: integer }
        content: { type: string }
        score: { type: number }
...
    AutoIngestRequest:
      type: object
      required: [source]
      properties:
        source:
          type: string
          description: 'Source identifier, e.g. "mos:army:11B" or "unit:marines:MALS-12".'
    AutoIngestSummary:
      type: object
      required: [source, total, added, existing, failed]
      properties:
        source: { type: string }
        total: { type: integer, description: Number of curated docs in the package. }
        added: { type: integer, description: Newly downloaded and ingested. }
        existing: { type: integer, description: Already present, skipped. }
        failed: { type: integer }
    AutoIngestJob:
      type: object
      required:
        [id, source, status, totalCount, addedCount, existingCount,
         failedCount, errorMessage, createdAt, updatedAt]
      properties:
        id: { type: string }
        source: { type: string }
        status: { type: string, description: '"running" | "done" | "failed"' }
        totalCount: { type: integer }
        addedCount: { type: integer }
        existingCount: { type: integer }
        failedCount: { type: integer }
        errorMessage: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AutoIngestStatusResponse:
      type: object
      required: [source, job]
      properties:
        source: { type: string }
        job:
          oneOf:
            - $ref: '#/components/schemas/AutoIngestJob'
            - type: 'null'
...
    DocumentPresetTagsUpdate:
      type: object
      required: [presetIds]
      properties:
        presetIds:
          type: array
          items: { type: string }