# lib/api-zod/src/generated/types/documentPresetTagsUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DoD AI Tool Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentPresetTagsUpdate {
  presetIds: string[];
}