# lib/api-zod/src/generated/types/documentSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DoD AI Tool Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentSummary {
  id: string;
  title: string;
  sourceFilename: string;
  mimeType: string;
  sizeBytes: number;
  charCount: number;
  chunkCount: number;
  status: string;
  /**
   * Stable identifier of the auto-ingest source that produced this doc, e.g. "mos:army:11B" or "unit:marines:MALS-12". Null for documents the user uploaded by hand.

   * @nullable
   */
  autoSource: string | null;
  /**
   * Original public URL the doc was downloaded from (auto-ingest only).
   * @nullable
   */
  sourceUrl: string | null;
  /** @nullable */
  errorMessage: string | null;
  uploadedAt: Date;
  /** @nullable */
  processedAt: Date | null;
  /** IDs of every mission preset this document is currently linked to. */
  presetIds: string[];
}